Eval Unit Tests for Adversarial Eval Testing (#82)
* add unit tests for input mod

Adds a unit test to check that a generated kernel which modifies the original inputs fails the correctness check.

For the square matmul problem, the kernel zeros out the inputs and returns a matrix of 0s. This will fail correctness/pass the test as long as the reference implementation is ran first. If we swap the order, the test will fail as the reference implementation will operate on tensors of 0s and it will look like the generated kernel computed the correct output.

* add unit test for reusing results

Adds a unit test to check that a generated kernel which attempts to access the result from the PyTorch reference model in memory fails the correctness check.

If a generated kernel uses empty_like, the CUDA caching allocator can re-use the physical memory of the previously computed result. All the kernel needs to do is return immediately and it will pass the correctness check.

Note that in order to reproduce this, we need to copy the PyTorch output to the CPU and delete the output object. Then empty_like will fetch the physical memory for the output object.

* test bhavesh's unit test

* use generic shape

use generic matmul shape for cache reuse adversarial kernel rather than requiring a square matmul.

* excessive speedup unit test via non default stream

* update timing signature

* update non default stream kernel

make a non-blocking non-default stream, and use cublasGemmEx rather than at::matmul:

* reduce trial for adverseiral stream hack

* show unrealistic speedup

* flag excessive speedups in eval script

eval script now flags excessive speedups by timing pytorch reference.

* reogranize a bit to flag timing on mian stream fails

* update EVAL.md with unit test summary

* ready for merge, update guide a bit (more for sahan to keep adding in other PRs)

---------

Co-authored-by: Simon Guo <simonguo@stanford.edu>